refactor(services) se modularon funciones y se integraron los primeros tests

+++ src/main/java/com/cesde/eventhub/entity/Order.java
@@ -51,7 +51,7 @@ public class Order {
     @Enumerated(EnumType.STRING)
     private OrderStatus status = OrderStatus.PENDIENTE;
 
-    private BigDecimal total; 
+    private double total; 
 
     @Enumerated(EnumType.STRING)
     private PaymentMethod paymentMethod;
